requirements.txt
colour==0.1.5
coverage==4.5.3
Pillow==6.0.0
